pyproject.toml
[tool.poetry]
name = "notion-autotests-api"
version = "0.1.0"
description = "Notion Autotest API"
authors = ["Sergey Galtsov <galtsovsergey@mail.ru>"]
readme = "README.md"
packages = [{include = "notion_autotests_api"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
pytest = "^7.2.2"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
allure-pytest = "^2.13.1"
python-dotenv = "^1.0.0"
curlify = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
